Wrap ref.watch based builder functions by Consumer
