+++ app/lib/features/chat/widgets/bubble_builder.dart
@@ -2,19 +2,19 @@ import 'package:acter/common/providers/common_providers.dart';
 import 'package:acter/common/themes/app_theme.dart';
 import 'package:acter/features/chat/providers/chat_providers.dart';
 import 'package:acter/features/chat/widgets/custom_message_builder.dart';
-import 'package:acter/features/chat/widgets/image_message_builder.dart';
-import 'package:acter/features/chat/widgets/text_message_builder.dart';
-import 'package:acter_avatar/acter_avatar.dart';
-import 'package:bubble/bubble.dart';
 import 'package:acter/features/chat/widgets/emoji_reaction_item.dart';
 import 'package:acter/features/chat/widgets/emoji_row.dart';
+import 'package:acter/features/chat/widgets/image_message_builder.dart';
+import 'package:acter/features/chat/widgets/text_message_builder.dart';
+import 'package:acter/features/home/providers/client_providers.dart';
+import 'package:acter_avatar/acter_avatar.dart';
+import 'package:bubble/bubble.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:swipe_to/swipe_to.dart';
 
 class BubbleBuilder extends ConsumerWidget {
-  final String userId;
   final Widget child;
   final types.Message message;
   final bool nextMessageInGroup;
@@ -25,16 +25,15 @@ class BubbleBuilder extends ConsumerWidget {
     required this.child,
     required this.message,
     required this.nextMessageInGroup,
-    required this.userId,
     required this.enlargeEmoji,
   }) : super(key: key);
 
-  bool isAuthor() {
-    return userId == message.author.id;
-  }
-
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    final client = ref.watch(clientProvider);
+    final myId = client!.userId().toString();
+    final isAuthor = (myId == message.author.id);
+
     final chatInputState = ref.watch(chatInputProvider);
     final chatInputNotifier = ref.watch(chatInputProvider.notifier);
     final chatRoomNotifier = ref.watch(chatRoomProvider.notifier);
@@ -47,7 +46,7 @@ class BubbleBuilder extends ConsumerWidget {
     return isMemberEvent
         ? child
         : SwipeTo(
-            onLeftSwipe: !isAuthor()
+            onLeftSwipe: !isAuthor
                 ? null
                 : () {
                     if (chatRoomNotifier.currentMessageId != null) {
@@ -62,7 +61,7 @@ class BubbleBuilder extends ConsumerWidget {
                       chatInputNotifier.setReplyWidget(child);
                     }
                   },
-            onRightSwipe: isAuthor()
+            onRightSwipe: isAuthor
                 ? null
                 : () {
                     if (chatInputState.emojiRowVisible) {
@@ -78,7 +77,6 @@ class BubbleBuilder extends ConsumerWidget {
                     }
                   },
             child: _ChatBubble(
-              isAuthor: isAuthor(),
               message: message,
               nextMessageInGroup: nextMessageInGroup,
               enlargeEmoji: enlargeEmoji,
@@ -89,14 +87,12 @@ class BubbleBuilder extends ConsumerWidget {
 }
 
 class _ChatBubble extends ConsumerWidget {
-  final bool isAuthor;
   final types.Message message;
   final bool nextMessageInGroup;
   final Widget child;
   final bool enlargeEmoji;
 
   const _ChatBubble({
-    required this.isAuthor,
     required this.message,
     required this.nextMessageInGroup,
     required this.child,
@@ -105,16 +101,17 @@ class _ChatBubble extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    final client = ref.watch(clientProvider);
+    final myId = client!.userId().toString();
+    final isAuthor = (myId == message.author.id);
+
     bool hasRepliedMessage = message.repliedMessage != null;
     return Column(
       crossAxisAlignment:
           isAuthor ? CrossAxisAlignment.end : CrossAxisAlignment.start,
       mainAxisSize: MainAxisSize.min,
       children: [
-        _EmojiRow(
-          isAuthor: isAuthor,
-          message: message,
-        ),
+        _EmojiRow(message: message),
         const SizedBox(height: 4),
         enlargeEmoji
             ? child
@@ -218,14 +215,15 @@ class _ChatBubble extends ConsumerWidget {
 }
 
 class _EmojiContainer extends StatefulWidget {
+  final bool isAuthor;
+  final types.Message message;
+  final bool nextMessageInGroup;
+
   const _EmojiContainer({
     required this.isAuthor,
     required this.message,
     required this.nextMessageInGroup,
   });
-  final bool isAuthor;
-  final types.Message message;
-  final bool nextMessageInGroup;
 
   @override
   State<_EmojiContainer> createState() => _EmojiContainerState();
@@ -330,7 +328,9 @@ class _EmojiContainerState extends State<_EmojiContainer>
     showModalBottomSheet(
       context: context,
       shape: const RoundedRectangleBorder(
-        borderRadius: BorderRadius.vertical(top: Radius.circular(15)),
+        borderRadius: BorderRadius.vertical(
+          top: Radius.circular(15),
+        ),
       ),
       isDismissible: true,
       builder: (BuildContext context) {
@@ -380,15 +380,16 @@ class _EmojiContainerState extends State<_EmojiContainer>
 }
 
 class _EmojiRow extends ConsumerWidget {
-  const _EmojiRow({
-    required this.isAuthor,
-    required this.message,
-  });
-  final bool isAuthor;
   final types.Message message;
+
+  const _EmojiRow({required this.message});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    final client = ref.watch(clientProvider);
+    final myId = client!.userId().toString();
+    final isAuthor = (myId == message.author.id);
+
     final chatRoomNotifier = ref.watch(chatRoomProvider.notifier);
     return Visibility(
       visible: message.id == chatRoomNotifier.currentMessageId,
@@ -414,24 +415,21 @@ class _EmojiRow extends ConsumerWidget {
 }
 
 class _OriginalMessageBuilder extends ConsumerWidget {
-  const _OriginalMessageBuilder({
-    required this.message,
-  });
-
   final types.Message message;
+
+  const _OriginalMessageBuilder({required this.message});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     if (message.repliedMessage is types.TextMessage) {
+      final w = message.repliedMessage!.metadata!['messageLength'] * 38.5;
       return TextMessageBuilder(
         message: message.repliedMessage as types.TextMessage,
-        messageWidth:
-            ((message.repliedMessage!.metadata!['messageLength']) * 38.5)
-                .toInt(),
+        messageWidth: w.toInt(),
         isReply: true,
       );
     } else if (message.repliedMessage is types.ImageMessage) {
-      var imageMsg = message.repliedMessage as types.ImageMessage;
+      final imageMsg = message.repliedMessage as types.ImageMessage;
       return Row(
         children: [
           Container(
@@ -469,11 +467,9 @@ class _OriginalMessageBuilder extends ConsumerWidget {
 }
 
 class _ReactionListing extends StatelessWidget {
-  const _ReactionListing({
-    required this.emojis,
-  });
-
   final List<String> emojis;
+
+  const _ReactionListing({required this.emojis});
 
   @override
   Widget build(BuildContext context) {
